Fixed Event Delegation Order
The Awake() methods that can be added to all Unity objects are called in a random order, so the C# events were not being passed to all item slots correctly. Changing the Awake() methods to Start() methods in Inventory.cs and Equipment.cs means that the events will always be passed along in the correct order, as Start() is called after Awake().